heroku用の環境設定

+++ database/migrations/2021_01_30_194341_change_password_column_line_friends_table.php
@@ -13,9 +13,11 @@ class ChangePasswordColumnLineFriendsTable extends Migration
      */
     public function up()
     {
-        Schema::table('line_friends', function (Blueprint $table) {
-            $table->dropColumn('password');
-        });
+        if (Schema::hasColumn('line_friends', 'password')) {
+            Schema::table('line_friends', function (Blueprint $table) {
+                $table->dropColumn('password');
+            });  
+        }
     }
 
     /**
